requirements.txt
pygame-ce
numpy
opencv-python
pillow
pywin32; sys_platform == "win32"
